point 3 and 4 solved
