+++ Phase1/src/test/java/SearchEngineTest.java
@@ -16,4 +16,18 @@ class SearchEngineTest {
         new SearchEngine("").run();
         assertEquals("no doc found\nno doc found\n",outputStream.toString().replaceAll("\r",""));
     }
+
+
+    @Test
+    public void searchEngineRun (){
+        ByteArrayInputStream in2 = new ByteArrayInputStream("cat\nnothing\nexit".getBytes());
+        System.setIn(in2);
+
+        ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
+        System.setOut(new PrintStream(outputStream));
+
+        SearchEngine searchEngine = new SearchEngine();
+        searchEngine.run();
+        assertEquals("no doc found\nno doc found\n",outputStream.toString().replaceAll("\r",""));
+    }
 }
